pyproject.toml
[tool.poetry]
name = "quiz-bot"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["Artem Nechai <lenikol95@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.12"
python-dotenv = "^0.20.0"
redis = "^4.3.3"
vk-api = "^11.9.8"


[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.16.1"
mypy = "^0.961"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
